first version with demand script working for new deployment scheme
