+++ cea/GUI/mobility_tool.py
@@ -1,5 +1,5 @@
 import cea
-from cea.interfaces.arcgis.toolbox import add_message
+from cea.interfaces.arcgis.CityEnergyAnalyst import add_message
 
 __author__ = "Daren Thomas"
 __copyright__ = "Copyright 2016, Architecture and Building Systems - ETH Zurich"
