+++ cea/GUI/radiation_tool.py
@@ -4,7 +4,7 @@ import cea
 import cea.globalvar
 import cea.inputlocator
 import cea.resources
-from cea.interfaces.arcgis.toolbox import add_message
+from cea.interfaces.arcgis.CityEnergyAnalyst import add_message
 
 __author__ = "Daren Thomas"
 __copyright__ = "Copyright 2016, Architecture and Building Systems - ETH Zurich"
