--- cea/interfaces/arcgis/City Energy Analyst.py
@@ -1,30 +1,0 @@
-import demand_tool
-
-__author__ = "Daren Thomas"
-__copyright__ = "Copyright 2016, Architecture and Building Systems - ETH Zurich"
-__credits__ = ["Daren Thomas"]
-__license__ = "MIT"
-__version__ = "0.1"
-__maintainer__ = "Daren Thomas"
-__email__ = "cea@arch.ethz.ch"
-__status__ = "Production"
-
-
-DemandTool = demand_tool.DemandTool
-# DataHelperTool = cea.GUI.data_helper_tool.DataHelperTool
-# EmissionsTool = cea.GUI.emissions_tool.EmissionsTool
-# EmbodiedEnergyTool = cea.GUI.embodied_energy_tool.EmbodiedEnergyTool
-# HeatmapsTool = cea.GUI.heatmaps_tool.HeatmapsTool
-# DemandGraphsTool = cea.GUI.demand_graphs_tool.DemandGraphsTool
-# RadiationTool = cea.GUI.radiation_tool.RadiationTool
-# ScenarioPlotsTool = cea.GUI.scenario_plots_tool.ScenarioPlotsTool
-# BenchmarkGraphsTool = cea.GUI.benchmark_graphs_tool.BenchmarkGraphsTool
-# MobilityTool = cea.GUI.mobility_tool.MobilityTool
-
-class Toolbox(object):
-    def __init__(self):
-        self.label = 'City Energy Analyst'
-        self.alias = 'cea'
-        # self.tools = [DataHelperTool, DemandTool, EmissionsTool, EmbodiedEnergyTool, HeatmapsTool, DemandGraphsTool,
-        #               RadiationTool, ScenarioPlotsTool, BenchmarkGraphsTool, MobilityTool]
-        self.tools = [DemandTool]
