--- cea/interfaces/arcgis/demand_tool.py
@@ -1,70 +1,0 @@
-import os
-
-import toolbox
-
-__author__ = "Daren Thomas"
-__copyright__ = "Copyright 2016, Architecture and Building Systems - ETH Zurich"
-__credits__ = ["Daren Thomas"]
-__license__ = "MIT"
-__version__ = "0.1"
-__maintainer__ = "Daren Thomas"
-__email__ = "cea@arch.ethz.ch"
-__status__ = "Production"
-
-
-class DemandTool(object):
-    """integrate the demand script with ArcGIS"""
-
-    def __init__(self):
-        self.label = 'Demand'
-        self.description = 'Calculate the Demand'
-        self.canRunInBackground = False
-
-    def getParameterInfo(self):
-        import arcpy
-        scenario_path = arcpy.Parameter(
-            displayName="Path to the scenario",
-            name="scenario_path",
-            datatype="DEFolder",
-            parameterType="Required",
-            direction="Input")
-        weather_name = arcpy.Parameter(
-            displayName="Weather file (choose from list or enter full path to .epw file)",
-            name="weather_name",
-            datatype="String",
-            parameterType="Required",
-            direction="Input")
-        weather_name.filter.list = toolbox.get_weather_names()
-
-        return [scenario_path, weather_name]
-
-    def isLicensed(self):
-        return True
-
-    def updateParameters(self, parameters):
-        return
-
-    def updateMessages(self, parameters):
-        scenario_path = parameters[0].valueAsText
-        if scenario_path is None:
-            return
-        if not os.path.exists(scenario_path):
-            parameters[0].setErrorMessage('Scenario folder not found: %s' % scenario_path)
-            return
-        if not os.path.exists(toolbox.get_radiation(scenario_path)):
-            parameters[0].setErrorMessage("No radiation data found for scenario. Run radiation script first.")
-        if not os.path.exists(toolbox.get_surface_properties(scenario_path)):
-            parameters[0].setErrorMessage("No radiation data found for scenario. Run radiation script first.")
-        return
-
-    def execute(self, parameters, _):
-        scenario_path = parameters[0].valueAsText
-        weather_name = parameters[1].valueAsText
-        if weather_name in toolbox.get_weather_names():
-            weather_path = toolbox.get_weather(weather_name)
-        elif os.path.exists(weather_name) and weather_name.endswith('.epw'):
-            weather_path = weather_name
-        else:
-            weather_path = toolbox.get_weather()
-
-        toolbox.run_cli(scenario_path, 'demand', '--weather', weather_path)
