--- cea/interfaces/arcgis/toolbox.py
@@ -1,102 +1,0 @@
-"""
-ArcGIS Tool classes for integrating the CEA with ArcGIS.
-
-These tools shell out to ``cli.py`` because the ArcGIS python version is old and we would like to decouple the
-python version used by CEA from the ArcGIS version.
-"""
-import os
-import subprocess
-import tempfile
-
-import exceptions
-
-__author__ = "Daren Thomas"
-__copyright__ = "Copyright 2016, Architecture and Building Systems - ETH Zurich"
-__credits__ = ["Daren Thomas"]
-__license__ = "MIT"
-__version__ = "0.1"
-__maintainer__ = "Daren Thomas"
-__email__ = "cea@arch.ethz.ch"
-__status__ = "Production"
-
-
-def add_message(msg, **kwargs):
-    import arcpy
-    """Log to arcpy.AddMessage() instead of print to STDOUT"""
-    if len(kwargs):
-        msg %= kwargs
-    arcpy.AddMessage(msg)
-    log_file = os.path.join(tempfile.gettempdir(), 'cea.log')
-    with open(log_file, 'a') as log:
-        log.write(msg)
-
-
-def get_weather_names():
-    """Shell out to cli.py and collect the list of weather files registered with the CEA"""
-    def get_weather_names_inner():
-        startupinfo = subprocess.STARTUPINFO()
-        startupinfo.dwFlags |= subprocess.STARTF_USESHOWWINDOW
-        command = [get_python_exe(), '-u', '-m', 'cea.cli', 'weather-files']
-        p = subprocess.Popen(command, stdout=subprocess.PIPE, startupinfo=startupinfo)
-        while True:
-            line = p.stdout.readline()
-            if line == '':
-                # end of input
-                break
-            yield line.rstrip()
-    return list(get_weather_names_inner())
-
-
-def get_weather(weather_name='default'):
-    """Shell out to cli.py and find the path to the weather file"""
-    return _cli_output('', 'weather-path', weather_name)
-
-
-def get_radiation(scenario_path):
-    """Shell out to cli.py and find the path to the ``radiation.csv`` file for the scenario."""
-    return _cli_output(scenario_path, 'locate', 'get_radiation')
-
-
-def get_surface_properties(scenario_path):
-    """Shell out to cli.py and find the path to the ``surface_properties.csv`` file for the scenario."""
-    return _cli_output(scenario_path, 'locate', 'get_surface_properties')
-
-
-def get_python_exe():
-    """Return the path to the python interpreter that was used to install CEA"""
-    try:
-        with open(os.path.expanduser('~/cea_python.pth'), 'r') as f:
-            python_exe = f.read().strip()
-            return python_exe
-    except:
-        raise exceptions.AssertionError("Could not find 'cea_python.pth' in home directory.")
-
-
-def _cli_output(scenario_path, *args):
-    """Run the CLI in a subprocess without showing windows and return the output as a string, whitespace
-    is stripped from the output"""
-    startupinfo = subprocess.STARTUPINFO()
-    startupinfo.dwFlags |= subprocess.STARTF_USESHOWWINDOW
-    result = subprocess.check_output(
-        [get_python_exe(), '-m', 'cea.cli', '--scenario', scenario_path] + list(args),
-        startupinfo=startupinfo)
-    return result.strip()
-
-
-def run_cli(scenario_path, *args):
-    """Run the CLI in a subprocess without showing windows"""
-    startupinfo = subprocess.STARTUPINFO()
-    startupinfo.dwFlags |= subprocess.STARTF_USESHOWWINDOW
-    process = subprocess.Popen(
-        [get_python_exe(), '-u', '-m', 'cea.cli', '--scenario', scenario_path] + list(args),
-        startupinfo=startupinfo,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE)
-    while True:
-        next_line = process.stdout.readline()
-        if next_line == '' and process.poll() is not None:
-            break
-        add_message(next_line.rstrip())
-    stdout, stderr = process.communicate()
-    add_message(stdout)
-    add_message(stderr)
